web-api/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-model-utils==4.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
